backend/requirements/base.txt
psycopg2-binary==2.8.6
sqlalchemy==2.0.*
alembic==1.11.*
eventsourcing-sqlalchemy==0.7
